# Conflicts:
#	engine/h2o/voxel_server/src/chunk_server.cpp

+++ engine/h2o/voxel/include/voxel/block.h
@@ -20,7 +20,7 @@ namespace h2o
 
     struct Block
     {
-        u64 id{};
+        u32 id{};
 
         Block() : Block(0) {}
         Block(u16 id) : id { id } {}
